リファクタリング（rubocop適用含む）

+++ app/controllers/items_controller.rb
@@ -37,12 +37,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if user_signed_in? && current_user.id == item.user.id
-      item.destroy
-      redirect_to root_path
-    else
-      redirect_to root_path
-    end
+    item.destroy if user_signed_in? && current_user.id == item.user.id
+    redirect_to root_path
   end
 
   private
